requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.1
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.48.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.59.0
googlesearch-python==1.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.4
Pillow==9.5.0
priority==2.0.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pyfiglet==0.8.post1
pyparsing==3.0.9
python-resize-image==1.1.20
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
termcolor==1.1.0
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
